refactor: rename core to config, nvim-mini org and cleanup
