+++ .config/nvim/init.lua
@@ -1,7 +1,12 @@
+-- enable lua module caching for faster startup
+if vim.loader then
+  vim.loader.enable()
+end
+
 -- load core configuration
-require('core.options')
-require('core.keymaps')
-require('core.autocmds')
+require('config.options')
+require('config.keymaps')
+require('config.autocmds')
 
 -- add mason bin to PATH
 vim.env.PATH = vim.fn.stdpath('data') .. '/mason/bin:' .. vim.env.PATH
